Camera (실제 Object 위치와 상대적인 Render 위치에 Camera를 일정시간동안 이동시키기)

+++ 2022_009_WinAPICore/Client/CAnimation.h
@@ -39,7 +39,7 @@ public:
 		m_iCurFrm = _iFrameIdx;
 		m_fAccTime = 0.f;
 	}
-	int GetMaxFrame() const { return m_vecFrm.size(); }
+	UINT GetMaxFrame() const { return (UINT)m_vecFrm.size(); }
 
 private:
 	void SetName(const wstring& _strName) { m_strName = _strName; }
